Use definition file for enumerations
This keeps the information about each enumeration value, its encoding,
and its name together. It is stored in `src/binary/defs.h`.

`src/binary/encoding.h` now has a `Decode` function for each
enumeration which converts from the binary format to the enumeration.
For some enums these are the same value (`ExternalKind` for example).
Keeping the enum separate from the encoding allows us to store
information about the enum value in a packed array.

The enums included are:

* `ValType`
* `ExternalKind`
* `Mutability`
* `SectionId` (this is new)
* `Opcode`

The old `Opcode` struct is now an enum. This way since we can use it an
a switch statement.