Remove Traits template param from types
This allowed choosing between `SpanU8` and `std::vector` for the
buffers; and `string_view` and `std::string` for the names. This was
originally added when using a callback-based interface, where the reader
was creating the structure directly. But now that we use an
iterator-based interface, it's not necessary to make this templatized.