Store locals in context in value type groups
Locals are stored in the binary format as count/value type pairs. This
makes it trivial to specify a huge number (2**31 say) of local
variables, so storing each local in a vector is too expensive to
validate.

Instead, we can store the partial sum of all local variables in one
vector, and a value type in another. For example, given the following
locals:

  i32 * 3, i64 * 6, f32 * 2, f64 * 4

We would store the following:

  partial sum:   3,   9,  11,  15
        types: i32, i64, f32, f64

You can then use `std::upper_bound` to search for any given index.