use StackType instead of ValueType in validation
The select instruction is polymorphic, and is required to pop and push
the correct type when the stack is in the unreachable state. Since any
type is valid in that case, there needs to be another value to push on
the type stack to represent this, e.g.

    unreachable
    i32.const 0
    select
    ;; stack contains one value of unknown type

This change adds a new `StackType` enum, which extends the `ValueType`
enum with the `StackType::Any` value. When checking types, `Any` will
match any other `ValueType`.