Change some validation failures to malformed binaries
Mismatching function/code, and data count/data section sizes are
malformed binaries, not validation failures. Similarly, out-of-order
sections are also malformed.

This commit changes fixes this. Some tests had to be changed as well
(because they relied on out-of-order sections).

The `lazy-module-utils` functions had a separate issue; since they will
cause a module to be read more than once, it was possible to get section
ordering errors even if the module was well-formed.