Always resolve when reading ScriptModule
This is required only because of the way resolving currently works; the
`NameMap`s are populated while reading, so starting to read a new module
will clear them.

Generally this is OK; if you are reading a script, you probably don't
care much about the internal names of a module. And you can still read a
`Module` directly which will not resolve names by default.

However, it would be better to separate this code out a bit, so
reading/resolving can be done separately if needed.