Add Desugar(Module&), which removes text sugar
The text format allows 3 different kinds of module-level sugar: inline
imports, inline exports, and inline segments. For each of these, the
`Desugar` function will remove the inline sugar and expand it into a new
`ModuleItem`.

This CL only implements `Desugar(Module&)`, and no other functions. We
may want to add these in the future, but for now it's simpler to
implement it on a whole `Module`. In particular, because desugaring adds
new module items and changes items into `Import`s.