[binary] Add Unpacked{Code,Expression} types
* Add `binary::UnpackedCode` and `binary::UnpackedExpression`
* Remove direct validation of `binary::Code`; it must be done piece by
  piece instead, since it requires reading `Instruction`s from
  `Expression`.
* Use `ValidateVisitor::OnInstruction` to validate instructions

This also fixes the bugs when validating instructions converted from
text; previously this would convert to binary and validate that way, but
would attempt to write the errors referencing the original text source,
but using the binary locations. This doesn't crash, but it does produce
bogus errors.

Now that the instructions are all unpacked, the text source locations
are bound directly.