Read and resolve ref types
A `ref` type can be one of the following:

    ht := func | extern | exn | <index>
    ref := ref null? <ht>

These can be used anywhere a reference type can be used. As a result,
the `Resolve()` pass is now expanded to handle the following:

* HeapType
* RefType
* ReferenceType
* ValueType
* ValueTypeList
* FunctionType
* BoundValueType
* BoundValueTypeList
* BoundFunctionType
* TypeEntry
* TableType
* TableDesc
* GlobalType
* GlobalDesc

A `TypeEntry` must be handled specially; it can both define and use a
type name, e.g.:

    (type $A (func))
    (type $B (func (param (ref $A)))

This is handled by adding an additional `DefineTypes()` pass, which
occurs before the `Define()` pass and collects all defined type names
before attempting to resolve them.