Add implicit `end` opcode when reading text
A function always has an `end` opcode, though it is never written in the
text format.

The final `end` instruction in a function is used by the validator to
determine whether the stack contents matches the function result. Since
this can fail, it should have a source location.

Previously, this implicit `end` instruction was added only when
converting from text to binary, but this means that there is no good
location to use for this new instruction. You can use the last
instruction in the function, but what if the function has no
instructions, like `(func)`?

This change implements a better solution, which uses the function's
final right parenthesis as the implicit end instruction. Now the
instruction always has a location, and the `ToBinary` pass doesn't need
to do anything special.