Validate type entries with defined type count
The validation algorithm can be done while streaming, so all information
must be gathered in binary order. Since the function-references proposal
introduces types that can refer to other types, it is possible to
reference a type that is not yet defined, e.g.

```
(type $t1 (func (result (ref $t2))))
(type $t2 (func (result (ref $t1))))
```

There's two potential ways to resolve this:

1) Defer validation of the type entries until all types have been read,
i.e. at the end of the type section. This has the drawback that
validation errors for a type entry don't occur when you would expect,
even if the types aren't self-referential. You could do further
inspection to determine whether they need to deferred, of course.

2) Take into account the number of expected type entries, which is
available at the start of the type section. This is only useful if the
only information you need is whether the index is valid. Fortunately, at
least for now, this is true.