Match equivalent function types
Implement EquivalentTypes class, which wraps a DisjointSet as well as a
set of current variable equivalence assumptions. When two variables are
shown to be equal, they can be added to the disjoint set. If they are
not equal then they need to need to be kept in the assumption map, since
the disjoint set doesn't have enough information to distinguish unequal
variables from variables that haven't been checked yet.