Fix validation of func.bind w/ polymorphic stack
The result type of func.bind is always known, so it needs to push this
type even when there is no function type given on the operand stack.

To test this, the `TestSignature` function is extended to make sure that
the result types are as expected. This is already true for nearly all
instructions, but for `call_ref` it's not possible to know the expected
result types with a polymorphic stack. So the `TestSignature` function
is split into multiple parts so `call_ref` can be handled differently.