Remove use of libfmt in wasp library
libfmt is a really nice library, but I don't think we want to take it as
a dependency for the library. It also seems to make compilation quite a
bit slower (maybe improved in future versions?). Part of the reason for
that, I think, is that user-defined formatter functions must be defined
as templates.

This CL uses `operator<<(std::ostream&, ...)` instead, which is well
known to be slower, but hopefully is not used enough in wasp to make
much of a difference.

The structure of the code isn't changed much, since the `fmt::format`
function has been replaced with `wasp::format`. Its behavior is
completely different:

```
    // prints "hello 13 world"
    wasp::format("hello ", 13, " world");
```

Since the behavior is so different, it probably should have a different
name. I'll do that in another CL.