Implement some of the match behavior for GC
This change is the start of what will be required for further
validation. It implements equality and subtype matching for the new rtt,
storage, field, struct, and array types, as well as folding these checks
into previous types.

I believe this requires creating a separate set to track relationships
between type indexes, one for equivalence and one for subtyping. I've
renamed `EquivalentTypes` to `TypeRelationSet` for that reason.