Use Abseil; remove fmt/span-lite/phmap
I tried using absl a while back and gave up on it. But now it actually
seems to be making things simpler!

TODO: The formatting functions still use C++ streams, which should be
replaced by the absl::str_format library. This means that wasp types
need to be wrapped in `concat()` when passing to an absl formatting
function.