Add ReadModule(), which reads full binary module
Previously the only way to read a module was lazily, section by section.
That's still the most efficient way to read a module, but if you know
that you'll need the full contents anyway, it's nicer to have a function
that reads all the sections into their components directly.

This change renames `ReadModule` to `ReadLazyModule`, and introduces a
new `ReadModule` function that reads into the existing `binary::Module`
struct.

This required a few other changes:

* The binary visitor was accidentally removing some of the locations;

* Some module reading errors are only logged to the `wasp::Errors`
  struct, so you can't easily know whether reading the module succeeded
  unless you query that struct. The `TextErrors` and `BinaryErrors`
  derived classes already had this ability (`has_error()`), so now it is a
  virtual method on `wasp::Errors` too (renamed to `HasError()`)