Fix spec test failure in func.wast (#34)
The wasm text format allows the use of a function type to be specified
one of three ways:

```
;; 1. As an explicit type index or symbol
(type 0)

;; 2. As params and results
(param i32) (result i64)

;; 3. Both; an explicit type index, params and results
(type 0) (param i32) (result i64)
```

Options 2 and 3 are text format sugar, since the binary format always
encodes these as an explicit type index.

Option 3 needs to check that the type at that index matches the
explicitly given params and results. Normally checking whether an index
is in bounds is a validation step, not a syntax checking step. However,
in this case if the index is out of bounds, then the types cannot be
checked, so it must be a syntax error.

The only additional complication is that there are two different kinds
of function type uses, bound and unbound. Bound function type uses allow
for the explicit params to be named.

Bound function type uses are seen in function definiitions:

```
(func (type 0) (param $x i32) ...)
```

Unbound function type uses are seen in instructions like
`call_indirect`:

```
call_indirect (type 0) (param i32)
```

Wasp models these differently, so they need to be resolved differently
too.