Fix a few function text writing bugs (#38)
The `end` instruction dedents the written output, but was assuming that
the instruction sequence was always well-formed. Now the `Dedent()`
function clamps to prevent underflow.

Also, the text format was assuming that a function's instruction
sequence would not have an `end` instruction, but the binary->text
conversion was including it. This means that an empty binary function
would be printed as `(func end)` in the text format, which is incorrect.
Now the text writing omits the final `end` instruction in the sequence,
if there is one.