Fix crash when validating invalid type index (#46)
As of the GC proposal, it is possible to have types that reference a
type index that has not yet been defined (i.e. recursive types). To
handle this, the validator will use the number of types declared
(`ValidCtx::defined_type_count`) in the type section to determine
whether the type index is valid.

The number of types that is actually defined (`ValidCtx::types::size()`)
may not match the number that was declared (e.g. if the type could not
be parsed, or is omitted). If this happens then accessing
`ValidCtx::types` with that index will crash.

This change updates `defined_type_count` at the end of the type section
to match the number of types that was actually defined. This way we can
always use `defined_type_count` and it will be consistent with the
number of defined types.

Also, we've added a few helper functions to make validating indexes
clearer: `ValidateTypeIndex`, `ValidateFunctionIndex`,
`ValidateMemoryIndex`, `ValidateTableIndex`, `ValidateGlobalIndex`,
`ValidateEventIndex`.