Fix IsMatch ordering (#58)
When checking for "sameness", the following is always true:
`IsSame(x,y) == IsSame(y,x)`. But the same is not true when matching,
since `IsMatch(x, y)` is not necessarily equal to `IsMatch(y, x)`. This
means that the `DisjointSet` cannot be used for `IsMatch`, since there's
no way to specify ordering. Instead, matching uses only the "assumption
map", where the map's key specifies the order.